# src/main/resources/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Financial Asset Tracking API",
    "description": "Financial Asset Information",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/v1/asset/{id}": {
      "get": {
        "tags": [
          "asset-detail-controller"
        ],
        "operationId": "getAssetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "asset-detail-controller"
        ],
        "operationId": "updateAsset",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Asset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "asset-detail-controller"
        ],
        "operationId": "deleteAssetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/asset": {
      "post": {
        "tags": [
          "asset-detail-controller"
        ],
        "operationId": "addAsset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Asset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assets": {
      "get": {
        "tags": [
          "asset-detail-controller"
        ],
        "operationId": "getAllAssets",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Asset"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/asset/name/{name}": {
      "get": {
        "tags": [
          "asset-detail-controller"
        ],
        "operationId": "getAssetByName",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "maxLength": 13,
              "minLength": 10,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Asset": {
        "required": [
          "costBasis",
          "name",
          "price",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "type": "string"
          },
          "quantity": {
            "maximum": 999999999,
            "minimum": 1,
            "type": "integer",
            "format": "int64"
          },
          "price": {
            "maximum": 9999999999.99,
            "exclusiveMaximum": false,
            "minimum": 0.00,
            "exclusiveMinimum": false,
            "type": "number"
          },
          "costBasis": {
            "maximum": 9999999999.99,
            "exclusiveMaximum": false,
            "minimum": 0.00,
            "exclusiveMinimum": false,
            "type": "number"
          },
          "notes": {
            "maxLength": 500,
            "minLength": 1,
            "type": "string"
          }
        }
      }
    }
  }
}